+++ Cargo.toml
@@ -4,10 +4,22 @@ resolver = "2"
 
 [workspace.dependencies]
 
+# workspace
+zingo_test_vectors = { path = "zingo_test_vectors" }
 
+# zcash
+zcash_protocol = { git = "https://github.com/zcash/librustzcash", rev = "d387aed7e04e881dbe30c6ff8b26a96c834c094b", features = [
+  "local-consensus",
+] }
+
+# zebra
+
+zebra-chain = "2.0"
+zebra-node-services = "1.0"
+zebra-rpc = "2.0"
+
+# other
 http = "1.1.0"
-
-
 getset = "0.1.3"
 hex = "0.4.3"
 json = "0.12.4"
@@ -18,10 +30,3 @@ thiserror = "1.0.64"
 tokio = "1.42.0"
 tracing = "0.1.40"
 tracing-subscriber = "0.3.15"
-zcash_protocol = { git = "https://github.com/zcash/librustzcash", rev = "d387aed7e04e881dbe30c6ff8b26a96c834c094b", features = [
-  "local-consensus",
-] }
-
-zebra-chain = "2.0"
-zebra-node-services = "1.0"
-zebra-rpc = "2.0"
